Fix usage of leeway
Authlib expects a relative leeway (a timedelta) instead of an absolute
point in time.

The checks in DRFIDToken have become redundant:
The exp header is already checked in the authlib package. The iat
property, according to the spec, can be used to verify the age of the
token. This does not have to compare to an absolute leeway.